refactor: extract hooks
